Add KmsClient for HC Vault.
This had been added in https://github.com/tink-crypto/tink/pull/712, but it was later removed because we thought that it is not needed. But there are some users that depend on these KMS client objects, so we add it now.

I did some edits to the original code, especially to the tests.

PiperOrigin-RevId: 636477645
Change-Id: I8039c021f2fb2cb2d1a608f3657a4f002afcf4ba